+++ pytest_inmanta/plugin.py
@@ -48,7 +48,6 @@ from inmanta import compiler, config, const, module, plugins, protocol
 from inmanta.agent import cache
 from inmanta.agent import config as inmanta_config
 from inmanta.agent import handler
-from inmanta.agent import io as agent_io
 from inmanta.agent.handler import HandlerContext, ResourceHandler
 from inmanta.const import ResourceState
 from inmanta.data import LogLine
@@ -73,7 +72,6 @@ if typing.TYPE_CHECKING:
     # Local type stub for mypy that works with both pytest < 7 and pytest >=7
     # https://docs.pytest.org/en/7.1.x/_modules/_pytest/legacypath.html#TempdirFactory
     import py
-    from inmanta.agent.io.local import IOBase
 
     class TempdirFactory:
         def mktemp(self, path: str) -> py.path.local:
@@ -1076,7 +1074,11 @@ class Project:
 
         c.open_version(resource.id.version)
         try:
-            p = handler.Commander.get_provider(c, agent, resource)  # type: ignore
+            # ISO8 and later no longer have the cache argument
+            try:
+                p = handler.Commander.get_provider(agent, resource)  # type: ignore
+            except TypeError:
+                p = handler.Commander.get_provider(c, agent, resource)  # type: ignore
             p.set_cache(c)
             p.get_file = lambda x: self.get_blob(x)  # type: ignore
             p.stat_file = lambda x: self.stat_blob(x)  # type: ignore
@@ -1316,14 +1318,6 @@ class Project:
         return DeployResultCollection(
             first_dryrun=first_dryrun, deploy=deploy, last_dryrun=last_dryrun
         )
-
-    def io(self, run_as_root: bool = False) -> "IOBase":
-        version = 1
-        if run_as_root:
-            ret = agent_io.get_io(None, "ssh://root@localhost", version)
-        else:
-            ret = agent_io.get_io(None, "local:", version)
-        return ret
 
     def create_module(self, name: str, initcf: str = "", initpy: str = "") -> None:
         module_dir = os.path.join(self._test_project_dir, "libs", name)