+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 2.11.0 (?)
 Changes in this release:
+- Remove io support to be compatible with ISO8
 - Add strict mode in `get_resource` method to  the `project` fixture and `Result` object (by default set to `False`). 
 
 ## Updating